requirements.txt
appnope==0.1.0
astroid==2.0.4
autopep8==1.4.3
backcall==0.1.0
decorator==4.3.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.14.5
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pygame==1.9.4
Pygments==2.2.0
pylint==2.1.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyserial==3.4
readline==6.2.4.1
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
wrapt==1.10.11
